cloud_build_codey/requirements.txt
google-cloud-aiplatform[preview]==1.32.0